add physical_constraint parser and fix some bug of tcl command read_constraint_file

+++ src/par/src/ChipletPartitioner.cpp
@@ -1,7 +1,7 @@
 #include "ChipletPartitioner.h"
 
 namespace par {
-    void ChipletPartitioner::initPhisicalConstraints(core_box core_box, int chiplet_area, int num_chiplets, 
+    void ChipletPartitioner::initPhisicalConstraints(core_box core_box, long int chiplet_area, int num_chiplets, 
         std::vector<utilization> chiplet_utilization, std::vector<aspect_ratio> chiplet_aspect_ratio) {
         _core_box = core_box;
         _chiplet_area = chiplet_area;
